backend/requirements.txt
fastapi[standard]
passlib
pydantic-settings
pyjwt
sqlmodel
websockets
textblob
asyncpg
alembic